refactor: Update get_phase_fractions function parameters in _funcs.py
The code changes in this commit update the get_phase_fractions function in the _funcs.py file. The function now accepts an additional parameter, paramsfile, which specifies the path to the parameters file. The default value is 'parameters.json'. The log parameter is also updated to have a default value of False.

This commit improves the functionality of the get_phase_fractions function by allowing for more flexibility in specifying the parameters file and controlling the logging behavior.

This commit message follows the established convention of using a prefix to indicate the type of change, followed by a concise and descriptive message.