refactor: Add endianess parameter to import_model and loadmat functions
The code changes in this commit add a new parameter, endianess, to the import_model and loadmat functions in the _funcs.py file. This parameter allows for specifying the endianess of the imported data, either 'big' or 'small'. This addition enhances the functionality of the input_output module by providing support for different endianess when importing model files.

This commit message follows the established convention of using a prefix to indicate the type of change, followed by a concise and descriptive message.